[update] change float64 to int64 in amount

+++ internal/models/ipl_rate.go
@@ -7,7 +7,7 @@ type IplRate struct {
 	RtID      string     `json:"rt_id" gorm:"not null"`
 	BlockID   *string    `json:"block_id" gorm:"null"`
 	HouseType *HouseType `json:"house_type" gorm:"null"`
-	Amount    float64    `json:"amount" gorm:"not null"`
+	Amount    int64      `json:"amount" gorm:"not null"`
 	StartDate time.Time  `json:"start_date" gorm:"not null"`
 	Rt        Rt         `gorm:"foreignKey:RtID" json:"rt"`
 	Block     Block      `gorm:"foreignKey:BlockID" json:"block"`
